backend/requirements.txt
fastapi
uvicorn
python-docx
pdfminer.six
pydantic
typing-extensions